merged

+++ src/utilities.cpp
@@ -56,7 +56,7 @@ void clearCIN(void) {
 //
 //Input a string to be returned to the calling function.
 //
-string inputString(string promptString, int minLength, int maxLength) {
+string inputString(string promptString, unsigned long minLength, unsigned long maxLength) {
     string userInput;
     
     while (true) {
@@ -109,7 +109,7 @@ int inputInt(string promptString, int minValue, int maxValue, int exitValue) {
 //Input a double value to be returned to the calling function.  Also, validate the
 //the minimum and maximum value of the double input.
 //
-double inputDouble(string promptString, double minValue, double maxValue, double exitValue) {
+double inputDouble(string promptString, unsigned long minValue, unsigned long maxValue, double exitValue) {
     double userInput;
     
     cout << promptString << " \n";
